Fix parsing the short optional arguments
This patch fixes parsing of the optional argument to the short options
`-l` (for `--list`) and `-w` (for `--download`) to accept the target
file/directory on the iOS device. The functionality was already
implemented and working with the long version of the options. That is:

* `-l /Library` did not work and is fixed now;

* `--list=/Library` used to and continues to work;

* `--list /Library` did not and does not work (`getopt_long` doesn't
support that).

The fix is to remove double colons after these short options. A few
unused one-letter options were also removed.