Update ios-deploy.m
The error code being sent for "Device Lock" here was getting sent as -402652958 rather than 3892314338.

Currently  check_error() casts the return to `unsigned int` so also casting the call for image-mounting were correctly sending the expected unsigned value in the JSON; this makes the JSON value for image-mounting issues come through as an unsigned int which can be compared to the hex values that are expected(e.g. 0xe80000e2 = 3892314338, not -402652958 ).